refactor rich messaging code
