+++ kommunicateui/src/main/java/com/applozic/mobicomkit/uiwidgets/conversation/richmessaging/webview/AlWebViewJsInterface.java
@@ -1,4 +1,4 @@
-package com.applozic.mobicomkit.uiwidgets.conversation.richmessaging.payment;
+package com.applozic.mobicomkit.uiwidgets.conversation.richmessaging.webview;
 
 import android.content.Context;
 import android.widget.Toast;
@@ -9,15 +9,15 @@ import com.applozic.mobicomkit.uiwidgets.R;
  * Created by ashish on 07/03/18.
  */
 
-public class PaymentJsInterface {
+public class AlWebViewJsInterface {
     Context mContext;
 
-    PaymentJsInterface(Context c) {
+    AlWebViewJsInterface(Context c) {
         mContext = c;
     }
 
     @android.webkit.JavascriptInterface
     public void success(long id, final String paymentId) {
-        Toast.makeText(mContext, mContext.getString(R.string.payment_successful), Toast.LENGTH_SHORT).show();
+        Toast.makeText(mContext, mContext.getString(R.string.form_action_success), Toast.LENGTH_SHORT).show();
     }
 }
