+++ kommunicateui/src/main/res/layout/km_activity_payment.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="com.applozic.mobicomkit.uiwidgets.conversation.richmessaging.payment.PaymentActivity">
+    tools:context="com.applozic.mobicomkit.uiwidgets.conversation.richmessaging.webview.AlWebViewActivity">
 
     <FrameLayout
         android:layout_width="match_parent"
